geohash support

+++ speciesgrids/index.py
@@ -5,6 +5,8 @@ import logging
 import os
 import shutil
 import h3pandas  # noqa: F401
+import geohash2
+from pyquadkey2.quadkey import QuadKey
 
 
 logger = logging.getLogger(__name__)
@@ -19,7 +21,8 @@ class Indexer:
             logger.info(f"Processing source {source}")
             source_output_path = os.path.join(self.temp_path, source)
             logger.info(f"Clearing output directory {source_output_path}")
-            shutil.rmtree(source_output_path)
+            if os.path.exists(source_output_path) and os.path.isdir(source_output_path):
+                shutil.rmtree(source_output_path)
             os.makedirs(source_output_path, exist_ok=False)
 
             for file in self.get_source_files(source_path):
@@ -30,20 +33,49 @@ class Indexer:
 
     def summarize_occurrences(self, df: pd.DataFrame) -> pd.DataFrame:
 
-        if self.aggregation["type"] == "h3":
-            return self.summarize_occurrences_h3(df)
-
-    def summarize_occurrences_h3(self, df: pd.DataFrame) -> pd.DataFrame:
-
         # fix types
 
         df["min_year"] = df["min_year"].astype("Int64")
         df["max_year"] = df["max_year"].astype("Int64")
 
+        # summarize
+
+        if self.aggregation["type"] == "h3":
+            return self.summarize_occurrences_h3(df)
+        elif self.aggregation["type"] == "geohash":
+            return self.summarize_occurrences_geohash(df)
+        else:
+            raise ValueError(f"Unsupported aggregation type {self.aggregation['type']}")
+
+    def summarize_occurrences_geohash(self, df: pd.DataFrame) -> pd.DataFrame:
+
+        # calculate geohash, drop coordinates
+
+        df["cell"] = df.apply(lambda row: geohash2.encode(row["decimalLatitude"], row["decimalLongitude"], precision=self.aggregation["resolution"]), axis=1)
+        df.drop(["decimalLongitude", "decimalLatitude"], axis=1, inplace=True)
+
+        # aggregate by geohash
+
+        df = df.groupby(["cell", "species", "AphiaID"], dropna=False).agg(
+            records=("records", lambda x: x.sum()),
+            min_year=("min_year", lambda x: x.min(skipna=True)),
+            max_year=("max_year", lambda x: x.max(skipna=True))
+        ).reset_index()
+
+        # calculate center geometry and quadkey, drop geometry
+
+        df[["y", "x"]] = df["cell"].apply(lambda x: geohash2.decode_exactly(x)[0:2]).apply(pd.Series)
+        df["quadkey"] = df.apply(lambda row: str(QuadKey.from_geo((row["y"], row["x"]), self.quadkey_level)), axis=1)
+        df = pd.DataFrame(df.drop(columns=["x", "y"]))
+        df.reset_index(inplace=True)
+
+        return df
+
+    def summarize_occurrences_h3(self, df: pd.DataFrame) -> pd.DataFrame:
+
         # calculate h3, drop coordinates
 
         df = df.h3.geo_to_h3(self.aggregation["resolution"], "decimalLatitude", "decimalLongitude", set_index=True)
-        # df.reset_index(inplace=True)
         df["cell"] = df.index
         df.drop(["decimalLongitude", "decimalLatitude"], axis=1, inplace=True)
 
@@ -68,8 +100,6 @@ class Indexer:
     def index_to_quadkey(self, source_file, output_path):
         """Indexes an occurrence parquet file to H3 and partitions output by quadkey."""
 
-        worms_output_path = os.path.join(self.temp_path, "worms_mapping.parquet")
-
         # create query
 
         res_cols = duckdb.query(f"describe select * from read_parquet('{source_file}')").fetchall()
@@ -86,7 +116,7 @@ class Indexer:
                     max(year) as max_year,
                     count(*) as records
                 from read_parquet('{source_file}')
-                left join read_parquet('{worms_output_path}') worms on specieskey = ID
+                left join read_parquet('{self.worms_output_path}') worms on specieskey = ID
                 where worms.AphiaID is not null and decimallongitude is not null and decimallatitude is not null
                 group by decimallongitude, decimallatitude, worms.scientificName, worms.AphiaID
             """
